Adds new video played, cleaning up
People on the KotB subreddit said there may be issues with the functionality of the Squire so I thought I'd check it out (having experience with JS). Only change I ended up making were to the alarm, which some people said was inaudible. Sorry for the slight reformat, it's from copying and pasting the code and shouldn't affect indentation.